Moved tests to match the code structure
The folder `integration/contracts` had tests for the proxies, to make
them easier to find they where moved to match the actual code structure.
