+++ raiden/tests/integration/network/proxies/test_discovery.py
