+++ raiden/tests/integration/network/proxies/test_payment_channel.py
@@ -6,7 +6,7 @@ from raiden.exceptions import ChannelOutdatedError
 from raiden.network.blockchain_service import BlockChainService
 from raiden.network.proxies import PaymentChannel, TokenNetwork
 from raiden.network.rpc.client import JSONRPCClient
-from raiden.tests.integration.contracts import BalanceProof
+from raiden.tests.integration.network.proxies import BalanceProof
 from raiden.utils import privatekey_to_address
 from raiden.utils.signer import LocalSigner
 from raiden_contracts.constants import TEST_SETTLE_TIMEOUT_MIN
