+++ raiden/tests/integration/network/proxies/test_secret_registry.py
