+++ raiden/tests/integration/network/proxies/test_token.py
