+++ raiden/tests/integration/network/proxies/test_token_network.py
@@ -13,7 +13,7 @@ from raiden.exceptions import (
 from raiden.network.blockchain_service import BlockChainService
 from raiden.network.proxies import TokenNetwork
 from raiden.network.rpc.client import JSONRPCClient
-from raiden.tests.integration.contracts import BalanceProof
+from raiden.tests.integration.network.proxies import BalanceProof
 from raiden.utils.signer import LocalSigner
 from raiden_contracts.constants import (
     TEST_SETTLE_TIMEOUT_MAX,
